Made the black-death bug go away
Before, the background would become all black after a while for some
reason. It looks like the reason was that I was creating new pens every
frame, which somehow made new pens be black/not work at all. Destroying
the pen at the end of each frame was exactly what was needed to stop this
from happening. It does not seem to be because of memory exhaustion,
probably just some inner limit to the number of pens / objects